feat : 질병 정보를 검색할수있는 검색창 구현

+++ src/components/SearchModal/components/SearchHistoryList.tsx
@@ -1,32 +1,41 @@
 import React from "react";
 import ListMagnifier from "../../Buttons/ListMagnifier";
 import RemoveButton from "../../Buttons/RemoveButton";
+import { useContext } from "react";
+import { SearchContext } from "../../../contexts/SearchContext";
 
 interface SearchHistoryListProps {
-  searches: string[];
   onClickItem: (search: string) => void;
   onDeleteItem: (search: string) => void;
 }
 
 const SearchHistoryList: React.FC<SearchHistoryListProps> = ({
-  searches,
   onClickItem,
   onDeleteItem,
-}) => (
-  <ul>
-    {searches.map((search, index) => (
-      <li
-        key={index}
-        onClick={() => onClickItem(search)}
-        className="mb-1 flex items-center hover:bg-gray-100 hover:cursor-pointer px-4 py-2 rounded"
-      >
-        <ListMagnifier />
-        <span className="text-lg text-gray-800">{search}</span>
+}) => {
+  const searchCtx = useContext(SearchContext);
 
-        <RemoveButton onClick={() => onDeleteItem(search)} />
-      </li>
-    ))}
-  </ul>
-);
+  if (!searchCtx) throw new Error("Cannot find search context");
+  const { controlHelper, searchHistory } = searchCtx;
+
+  return (
+    <ul>
+      {searchHistory.map((search, index) => (
+        <li
+          key={index}
+          onClick={() => onClickItem(search)}
+          className={`mb-1 flex items-center hover:bg-gray-100 hover:pointer px-4 py-2 rounded ${
+            index === controlHelper ? "bg-gray-200" : ""
+          }`}
+        >
+          <ListMagnifier />
+          <span className="text-lg text-gray-800">{search}</span>
+
+          <RemoveButton onClick={() => onDeleteItem(search)} />
+        </li>
+      ))}
+    </ul>
+  );
+};
 
 export default SearchHistoryList;
